Digital Library (PDF & Audio Books)
This project is a web application for downloading PDF and audio books, supported by a database. It is developed using the MVC architecture within the ASP.NET Core framework. Due to the large size of the source code, the project has been provided as a recorded video demonstration instead of uploading the full solution to the repository.